system-requirements.txt
# System Requirements (Debian/Ubuntu/Raspbian)
# Install via: sudo apt install $(grep -v '^#' system-requirements.txt | tr '\n' ' ')

# Python and build tools
python3-venv
python3-dev
python3-pip
build-essential
cmake
pkg-config
ninja-build
python-is-python3

# Python packages (system-installed)
python3-pil
python3-numpy
python3-opencv
python3-psutil
python3-websockets
python3-opengl
python3-moderngl

# Image and media libraries
libwebp-dev
libjpeg-dev
libturbojpeg0

# OpenGL and graphics libraries
libgl1-mesa-dev
libglu1-mesa-dev
libegl1-mesa-dev
libgles-dev
libglvnd-dev
libglfw3-dev
mesa-utils
mesa-utils-extra

# Audio and input libraries
libsdl2-dev
libasound2-dev

# System utilities
chrony
dnsutils
